Add price component
