+++ src/components/Product.js
@@ -2,32 +2,10 @@ import Link from 'next/link';
 import AddToCartButton from '../components/cart/AddToCartButton';
 import clientConfig from '../../client-config';
 import { isEmpty } from 'lodash';
+import Price from "./single-product/price";
 
 const Product = ( props ) => {
 	const { product } = props;
-
-	/**
-	 * Get discount percent.
-	 *
-	 * @param {String} regularPrice
-	 * @param {String} salesPrice
-	 */
-	const discountPercent = ( regularPrice, salesPrice ) => {
-		if( isEmpty( regularPrice ) || isEmpty(salesPrice) ) {
-			return null;
-		}
-
-		const formattedRegularPrice = parseInt( regularPrice?.substring(1) );
-		const formattedSalesPrice = parseInt( salesPrice?.substring(1) );
-		const discountPercent = ( formattedSalesPrice / formattedRegularPrice ) * 100;
-
-		return {
-			discountPercent: formattedSalesPrice !== formattedRegularPrice ? `(${discountPercent.toFixed(2)}%) OFF` : null,
-			strikeThroughClass: formattedSalesPrice < formattedRegularPrice ? 'product-regular-price mr-2 line-through text-sm text-gray-600 font-normal' : ''
-		}
-	}
-
-	const productMeta = discountPercent( product?.regularPrice, product?.price );
 
 	return (
 		// @TODO Need to handle Group products differently.
@@ -53,16 +31,7 @@ const Product = ( props ) => {
 						{ product.name ? product.name : '' }
 					</h3>
 					<div className="product-description text-sm text-gray-700" dangerouslySetInnerHTML={{ __html: (product?.description)}}/>
-						<h6 className="product-price text-gray-800 font-semibold mr-3 mb-5">
-							{/* Regular price */}
-						{ productMeta?.discountPercent ? <span className="product-price mr-2">{product?.regularPrice}</span> : null }
-
-						{/* Discounted price */}
-						<span className={productMeta?.strikeThroughClass}>{ product.price }</span>
-
-						{/* Discount percent */}
-						<span className="product-discount text-red-500 text-sm font-normal">{productMeta?.discountPercent}</span>
-						</h6>
+					<Price salesPrice={product?.price} regularPrice={product?.regularPrice}/>
 					<AddToCartButton product={ product }/>
 				</div>
